feat: refactor editor code
- Make the editor code reusable across multiple editor types
- This will be used for the starter code feature in a later PR
- Rename `Editor` to `RoomEditor` to indicate that its a code editor in the context of a given room
